+++ ios/CleverTapReact/CleverTapReact.mm
@@ -934,9 +934,13 @@ RCT_EXPORT_METHOD(suspendInAppNotifications) {
     [[self cleverTapInstance] suspendInAppNotifications];
 }
 
-RCT_EXPORT_METHOD(discardInAppNotifications) {
-    RCTLogInfo(@"[CleverTap discardInAppNotifications");
-    [[self cleverTapInstance] discardInAppNotifications];
+RCT_EXPORT_METHOD(discardInAppNotifications:(NSNumber *)dismissInAppIfVisible) {
+    RCTLogInfo(@"[CleverTap discardInAppNotifications: %@]", dismissInAppIfVisible);
+    if (dismissInAppIfVisible != nil) {
+        [[self cleverTapInstance] discardInAppNotifications:[dismissInAppIfVisible boolValue]];
+    } else {
+        [[self cleverTapInstance] discardInAppNotifications];
+    }
 }
 
 RCT_EXPORT_METHOD(resumeInAppNotifications) {