+++ src/index.js
@@ -12,7 +12,7 @@ const EventEmitter = Platform.select({
 * @param {int} libVersion - The updated library version. If current version is 1.1.0 then pass as 10100  
 */
 const libName = 'React-Native';
-const libVersion = 30700;
+const libVersion = 30800;
 CleverTapReact.setLibrary(libName,libVersion);
 
 function defaultCallback(method, err, res) {
@@ -937,9 +937,10 @@ var CleverTap = {
      * Suspends the display of InApp Notifications and discards any new InApp Notifications to be shown
      * after this method is called.
      * The InApp Notifications will be displayed only once resumeInAppNotifications() is called.
+     * @param {boolean} dismissInAppIfVisible - Optional. If true, dismisses the currently visible InApp notification.
      */
-    discardInAppNotifications: function () {
-        CleverTapReact.discardInAppNotifications();
+    discardInAppNotifications: function (dismissInAppIfVisible) {
+        CleverTapReact.discardInAppNotifications(dismissInAppIfVisible);
     },
 
     /**